+++ example/app/app.module.ts
@@ -7,14 +7,14 @@ import { DatePickerNgmodel } from './date-picker-ngmodel';
 import { DatePickerReactiveForms } from './date-picker-reactive-forms';
 import { DatePickerDivHostElement } from './date-picker-div-host-element';
 import { DatePickerInline} from './date-picker-inline';
-import { AngularMyDatePickerModule } from '../../projects/angular-mydatepicker/src/public-api';
+import { AngularMyDatePickerDirective } from '../../projects/angular-mydatepicker/src/public-api';
 
 @NgModule({
   declarations: [
     AppComponent, DatePickerNgmodel, DatePickerReactiveForms, DatePickerDivHostElement, DatePickerInline
   ],
   imports: [
-    BrowserModule, ReactiveFormsModule, FormsModule, AngularMyDatePickerModule
+    BrowserModule, ReactiveFormsModule, FormsModule, AngularMyDatePickerDirective
   ],
   providers: [],
   bootstrap: [AppComponent]