+++ projects/angular-mydatepicker/src/lib/components/selection-bar/selection-bar.component.ts
@@ -1,4 +1,5 @@
 import {Component, EventEmitter, Input, OnChanges, Output, ViewEncapsulation, SimpleChanges} from "@angular/core";
+import {CommonModule} from "@angular/common";
 import {IMyMonth} from "../../interfaces/my-month.interface";
 import {IMyOptions} from "../../interfaces/my-options.interface";
 import {OPTS, YEARS_DURATION, VISIBLE_MONTH, SELECT_MONTH, SELECT_YEAR, PREV_VIEW_DISABLED, NEXT_VIEW_DISABLED} from "../../constants/constants"
@@ -7,7 +8,7 @@ import {OPTS, YEARS_DURATION, VISIBLE_MONTH, SELECT_MONTH, SELECT_YEAR, PREV_VIE
     selector: "lib-selection-bar",
     templateUrl: "./selection-bar.component.html",
     encapsulation: ViewEncapsulation.None,
-    standalone: false
+    imports: [CommonModule]
 })
 export class SelectionBarComponent implements OnChanges {
   @Input() opts: IMyOptions;